#SO_UI-17: Created Router Wrapper and locked down routes for only logged-in users
