+++ public/style.css
@@ -36,42 +36,6 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
-.login-wrapper {
-  margin-top: 80px;
-  margin-bottom: 80px;
-}
-.form-login {
-  max-width: 380px;
-  padding: 15px 35px 45px;
-  margin: 0 auto;
-  background-color: #fff;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-}
-.form-login .form-login-heading,
-.form-login .checkbox-label {
-  margin-bottom: 30px;
-  font-weight: normal;
-}
-.form-login .form-control-input {
-  position: relative;
-  font-size: 16px;
-  height: auto;
-  padding: 10px;
-  @include box-sizing(border-box);
-}
-.form-login .form-control-input:focus {
-  z-index: 2;
-}
-.form-login input[type="text"] {
-  margin-bottom: -1px;
-  border-bottom-left-radius: 0;
-  border-bottom-right-radius: 0;
-}
-.form-login input[type="password"] {
-  margin-bottom: 20px;
-  border-top-left-radius: 0;
-  border-top-right-radius: 0;
-}
 .login-center {
   font-family: 'Raleway', sans-serif;
   -webkit-background-size: cover;
