+++ src/routes/AuthenticatedRoute.js
@@ -1,0 +1,29 @@
+import React, { Component } from 'react';
+import {
+    Route,
+    Redirect
+} from 'react-router-dom';
+
+import AuthService from "../containers/Login/AuthService";
+
+/**
+ * We have to make our own Authenticated Route.... So here it is -_-
+ * @param Component
+ * @param rest
+ * @returns {*}
+ * @constructor
+ */
+const AuthenticatedRoute = ({ component: Component, ...rest }) => (
+    <Route {...rest} render={props => (
+        AuthService.isAuthorized() ? (
+            <Component {...props}/>
+        ) : (
+            <Redirect to={{
+                pathname: '/login',
+                state: { from: props.location }
+            }}/>
+        )
+    )}/>
+)
+
+export default AuthenticatedRoute;
