+++ style/style.less
@@ -2,6 +2,5 @@
 @import './notFound.less';
 @import './app.less';
 @import './index.less';
-@import './login.less';
 
 @import './loginContainer.less';
